Added motion tracking data
Here's how this data is formatted:

The time parameter represents the time in the video in frames, starting at zero. The video is 30 frames/second, so a value of 30 represents one second in, 60 represents 2 seconds in, etc.

The coordinates represent the offset of the face from the center of the screen in pixels. For the y axis, larger numbers represet going up. The video is 1920x1080, so for example, the top-right corner would be at (960, 540) and the bottom-right corner would be at (960, -540).